agents/ralph_loop.py
from google.adk.agents import LoopAgent, Agent
from tools.policy_engine import global_policy
import json
import os

class RalphLoopManager:
    """자율 작업 완수 루프 및 무중단 복원력(ZDR) 관리"""

    def __init__(self, state_file: str = ".agent_state.json"):
        self.state_file = state_file

    def save_state(self, state: dict):
        """상태 영속화 (Persistence)"""
        with open(self.state_file, 'w') as f:
            json.dump(state, f)

    def load_state(self) -> dict:
        """상태 재구성 (Rehydrate / ZDR)"""
        if os.path.exists(self.state_file):
            with open(self.state_file, 'r') as f:
                return json.load(f)
        return {}

# 1. 자율 루프용 워커 정의
worker_agent = Agent(
    name="autonomous_worker",
    model="gemini-3-flash-preview",
    instruction="상태 파일의 TODO를 하나씩 해결하고 결과를 기록하세요.",
    # max_iterations guardrail은 LoopAgent에서 설정
)

# 2. Ralph Loop 구성
autonomous_loop = LoopAgent(
    name="ralph_loop",
    sub_agents=[worker_agent],
    max_iterations=10, # 엔터프라이즈 비용 통제 Guardrail
)

if __name__ == "__main__":
    manager = RalphLoopManager()
    current_state = manager.load_state()
    
    print("Starting Autonomous Ralph Loop with ZDR...")
    # LoopAgent 실행 로직 (ADK Runner 연동 필요)
