tools/mcp_client.py
import asyncio
from mcp import ClientSession
from mcp.client.sse import sse_client
from mcp.client.stdio import stdio_client, StdioServerParameters

class RemoteMCPIntegration:
    """엔터프라이즈 MCP 서버 연동 클래스 (Remote MCP - SSE 지향)"""
    
    def __init__(self, server_url: str = "https://onemcp.internal.example.com/sse"):
        """
        엔터프라이즈 환경에서는 Local MCP(stdio)보다, 
        조직 내 인증(OAuth) 및 Rate Limit 관리가 일원화된 Remote MCP(SSE)를 권장합니다.
        """
        self.server_url = server_url

    async def call_remote_tool(self, tool_name: str, arguments: dict):
        """Remote MCP 도구 호출 및 결과 반환 (A2A 연동 가능)"""
        print(f"🔗 Connecting to Remote MCP Server: {self.server_url}")
        
        # 실제 환경에서는 headers에 인증 토큰(Bearer)을 포함합니다.
        async with sse_client(self.server_url) as (read, write):
            async with ClientSession(read, write) as session:
                await session.initialize()
                result = await session.call_tool(tool_name, arguments)
                return result

if __name__ == "__main__":
    # Remote MCP 통합 테스트 예시 (URL은 가상)
    mcp = RemoteMCPIntegration()
    # asyncio.run(mcp.call_remote_tool("query_bigquery", {"dataset": "analytics"}))
    print("Remote MCP Client Initialized.")
